rename folder MainProject into TagCloudLayouters
